Moved MMIO into its own library
