+++ clem_mem.h
@@ -44,18 +44,6 @@ void clem_mem_create_page_mapping(struct ClemensMemoryPageInfo *page, uint8_t pa
 void clem_read(ClemensMachine *clem, uint8_t *data, uint16_t adr, uint8_t bank, uint8_t flags);
 void clem_write(ClemensMachine *clem, uint8_t data, uint16_t adr, uint8_t bank, uint8_t flags);
 
-void clem_mmio_reset(ClemensMMIO *mmio, clem_clocks_duration_t mega2_clocks_step);
-
-void clem_mmio_init(ClemensMMIO *mmio, struct ClemensMemoryPageMap **bank_page_map,
-                    clem_clocks_duration_t mega2_clocks_step, void *slot_expansion_rom);
-
-uint8_t clem_mmio_read(ClemensMMIO *mmio, struct ClemensTimeSpec *tspec, uint16_t addr,
-                       uint8_t flags, bool *mega2_access);
-void clem_mmio_write(ClemensMMIO *mmio, struct ClemensTimeSpec *tspec, uint8_t data, uint16_t addr,
-                     uint8_t flags, bool *mega2_access);
-
-void clem_mmio_restore(ClemensMMIO *mmio);
-
 #ifdef __cplusplus
 }
 #endif
