+++ serializer.c
@@ -86,10 +86,6 @@ struct ClemensSerializerRecord kTimer[] = {
     CLEM_SERIALIZER_RECORD_EMPTY()};
 
 struct ClemensSerializerRecord kDebugger[] = {
-    CLEM_SERIALIZER_RECORD_ARRAY(struct ClemensDeviceDebugger, kClemensSerializerTypeUInt32,
-                                 ioreg_read_ctr, 256, 0),
-    CLEM_SERIALIZER_RECORD_ARRAY(struct ClemensDeviceDebugger, kClemensSerializerTypeUInt32,
-                                 ioreg_write_ctr, 256, 0),
     CLEM_SERIALIZER_RECORD_UINT16(struct ClemensDeviceDebugger, pc),
     CLEM_SERIALIZER_RECORD_UINT8(struct ClemensDeviceDebugger, pc), CLEM_SERIALIZER_RECORD_EMPTY()};
 
@@ -236,7 +232,6 @@ struct ClemensSerializerRecord kMMIO[] = {
     CLEM_SERIALIZER_RECORD_OBJECT(ClemensMMIO, dev_rtc, struct ClemensDeviceRTC, kRTC),
     CLEM_SERIALIZER_RECORD_OBJECT(ClemensMMIO, dev_adb, struct ClemensDeviceADB, kADB),
     CLEM_SERIALIZER_RECORD_OBJECT(ClemensMMIO, dev_timer, struct ClemensDeviceTimer, kTimer),
-    CLEM_SERIALIZER_RECORD_OBJECT(ClemensMMIO, dev_debug, struct ClemensDeviceDebugger, kDebugger),
     CLEM_SERIALIZER_RECORD_OBJECT(ClemensMMIO, dev_audio, struct ClemensDeviceAudio, kAudio),
     CLEM_SERIALIZER_RECORD_OBJECT(ClemensMMIO, dev_iwm, struct ClemensDeviceIWM, kIWM),
     CLEM_SERIALIZER_RECORD_OBJECT(ClemensMMIO, dev_scc, struct ClemensDeviceSCC, kSCC),
@@ -311,7 +306,10 @@ struct ClemensSerializerRecord kMachine[] = {
     CLEM_SERIALIZER_RECORD_OBJECT(ClemensMachine, cpu, struct Clemens65C816, kCPU),
     CLEM_SERIALIZER_RECORD_OBJECT(ClemensMachine, tspec, struct ClemensTimeSpec, kTimeSpec),
     CLEM_SERIALIZER_RECORD_OBJECT(ClemensMachine, mem, struct ClemensMemory, kMemory),
-    CLEM_SERIALIZER_RECORD_INT32(ClemensMachine, resb_counter), CLEM_SERIALIZER_RECORD_EMPTY()};
+    CLEM_SERIALIZER_RECORD_INT32(ClemensMachine, resb_counter),
+    CLEM_SERIALIZER_RECORD_OBJECT(ClemensMachine, dev_debug, struct ClemensDeviceDebugger,
+                                  kDebugger),
+    CLEM_SERIALIZER_RECORD_EMPTY()};
 
 // see clem_disk.h
 struct ClemensSerializerRecord kNibbleDisk[] = {
